translations, renaming files, linted, theme colors
