+++ src/components/CreateCourse/Modal/CreateCourseModal.tsx
@@ -4,10 +4,10 @@ import { useTranslation } from 'react-i18next'
 import { Box, Button, Fab, Grid, Modal } from '@common/components'
 import { Close } from '@common/icons'
 import { RemoteCourse, User } from '@core'
+import { postCourse } from '@services'
 import { usePersistedStore } from '@store'
-import { postCourse } from '../../../services/Course/postCourse'
-import TableCourse from '../Table/TableCourse'
-import TableCourseDetails from '../Table/TableCourseDetails'
+import CreateCourseDetailsTable from '../Table/CreateCourseDetailsTable'
+import CreateCourseTable from '../Table/CreateCourseTable'
 
 type CourseModalProps = {
   open?: boolean
@@ -16,11 +16,13 @@ type CourseModalProps = {
 
 const CreateCourseModal = memo(({ open = false, handleClose }: CourseModalProps) => {
   const { t } = useTranslation()
+
   const [selectedCourse, setSelectedCourse] = useState<RemoteCourse>()
   const [activeStep, setActiveStep] = useState<number>(0)
-  const [value, setValue] = useState<Dayjs | null>(dayjs(new Date()))
-  const getUser = usePersistedStore((state) => state.getUser)
+  const [startDateValue, setStartDateValue] = useState<Dayjs | null>(dayjs(new Date()))
   const [user, setUser] = useState<User>()
+
+  const getUser = usePersistedStore((state) => state.getUser)
 
   const handleCourseSelection = (course: RemoteCourse) => {
     setSelectedCourse(course)
@@ -42,7 +44,7 @@ const CreateCourseModal = memo(({ open = false, handleClose }: CourseModalProps)
     const createCourse = {
       lms_id: selectedCourse?.id || 1,
       name: selectedCourse?.fullname || 'Test',
-      start_date: value ? formatDate(value.toDate()) : formatDate(new Date()),
+      start_date: startDateValue ? formatDate(startDateValue.toDate()) : formatDate(new Date()),
       university: user?.university || 'Test',
       created_by: user?.settings.user_id || 1,
       created_at: formatDate(new Date())
@@ -69,7 +71,7 @@ const CreateCourseModal = memo(({ open = false, handleClose }: CourseModalProps)
           }}>
           <Fab
             color="primary"
-            data-testid={'QuestionnaireResultsCloseButton'}
+            id={'create-course-modal-close-button'}
             onClick={handleClose}
             sx={{
               position: 'sticky',
@@ -80,7 +82,10 @@ const CreateCourseModal = memo(({ open = false, handleClose }: CourseModalProps)
           </Fab>
           {activeStep === 0 ? (
             <Grid>
-              <TableCourse onCourseSelect={handleCourseSelection} />
+              <CreateCourseTable
+                onCourseSelect={handleCourseSelection}
+                selectedCourseName={selectedCourse?.fullname || ''}
+              />
               <Grid container justifyContent="center" alignItems="center" sx={{ mt: 2 }}>
                 <Button id="add-course-button" variant="contained" color="primary" onClick={() => setActiveStep(1)}>
                   {t('components.CourseModal.selectCourse')}
@@ -89,7 +94,11 @@ const CreateCourseModal = memo(({ open = false, handleClose }: CourseModalProps)
             </Grid>
           ) : (
             <Grid>
-              <TableCourseDetails course={selectedCourse} datePickerValue={value} setDatePickerValue={setValue} />
+              <CreateCourseDetailsTable
+                remoteCourse={selectedCourse}
+                datePickerValue={startDateValue}
+                setDatePickerValue={setStartDateValue}
+              />
               <Grid container justifyContent="space-between" alignItems="center" sx={{ mt: 2 }}>
                 <Button id="add-course-button" variant="contained" color="primary" onClick={() => setActiveStep(0)}>
                   {t('appGlobal.back')}
