+++ src/components/CreateCourse/Table/CreateCourseDetailsTable.tsx
@@ -2,47 +2,34 @@ import { renderTimeViewClock } from '@mui/x-date-pickers'
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs'
 import { DateTimePicker } from '@mui/x-date-pickers/DateTimePicker'
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider'
-import { DemoContainer } from '@mui/x-date-pickers/internals/demo'
-import dayjs, { Dayjs } from 'dayjs'
-import { useCallback, useEffect, useState } from 'react'
+import { Dayjs } from 'dayjs'
 import { useTranslation } from 'react-i18next'
-import {
-  Divider,
-  Fab,
-  FormControl,
-  Grid,
-  InputLabel,
-  MenuItem,
-  Radio,
-  RadioGroup,
-  Select,
-  Stack,
-  TextField,
-  Typography
-} from '@common/components'
-import { Close } from '@common/icons'
+import { Grid, TextField } from '@common/components'
 import { RemoteCourse } from '@core'
 
 type TableCourseDetailsProps = {
-  course: RemoteCourse | undefined
+  remoteCourse: RemoteCourse | undefined
   datePickerValue: Dayjs | null
   setDatePickerValue: (value: Dayjs | null) => void
 }
 
-const TableCourseDetails = ({ course, datePickerValue, setDatePickerValue }: TableCourseDetailsProps) => {
-  const [selected, setSelected] = useState(0)
-  const [teacherselection, setteacherselection] = useState(0)
+const CreateCourseDetailsTable = ({ remoteCourse, datePickerValue, setDatePickerValue }: TableCourseDetailsProps) => {
   const { t } = useTranslation()
 
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs}>
       <Grid container direction="column" spacing={3}>
         <Grid item container direction="row" justifyContent="space-between" alignItems="center">
-          <TextField label="Kursname" defaultValue={course?.fullname} sx={{ width: '100%', mt: '1rem' }} required />
+          <TextField
+            label="Kursname"
+            defaultValue={remoteCourse?.fullname}
+            sx={{ width: '100%', mt: '1rem' }}
+            required
+          />
         </Grid>
         <Grid item container direction="row" justifyContent="space-between" alignItems="center">
           <DateTimePicker
-            label="Startdatum"
+            label={t('components.CreateCourseDetailsTable.startdate')}
             value={datePickerValue}
             onChange={setDatePickerValue}
             sx={{ width: '100%' }}
@@ -59,7 +46,6 @@ const TableCourseDetails = ({ course, datePickerValue, setDatePickerValue }: Tab
     </LocalizationProvider>
   )
 }
-
 // eslint-disable-next-line immutable/no-mutation
-TableCourseDetails.displayName = 'TableCourseDetails'
-export default TableCourseDetails
+CreateCourseDetailsTable.displayName = 'TableCourseDetails'
+export default CreateCourseDetailsTable
