+++ src/services/index.ts
@@ -19,6 +19,7 @@ export * from './SnackbarContext'
 export { SnackbarProvider } from './SnackbarProvider'
 export * from './connection'
 export { fetchCourses } from './Courses'
+export { postCourse } from './Course'
 export * from './xAPI'
 export { fetchILS, fetchListK, postILS, postListK } from './Questionnaire'
 export * from './debounce'
